requirements.txt
marshmallow
google-cloud-bigquery
protobuf
google-cloud-pubsub
google-cloud-logging
nltk
